made a new file and transferred the files there yo
